requirements.txt
Flask
reportlab
supabase-py